requirements.txt
numpy
cvzone
ultralytics
hydra-core
matplotlib
opencv-python
Pillow
PyYAML
requests
scipy
torch
torchvision
tqdm
filterpy
scikit-image
lap
deep-sort-realtime